Move more getters into ts_ast dir
