+++ src/compiler/typescript_to_rust_redesign/azle_generate/src/generators/canister_methods/functions.rs
@@ -88,33 +88,12 @@ pub fn generate_params_token_stream(names: &Vec<Ident>, types: &Vec<RustType>) -
 }
 
 fn generate_return_type(ast_fnc_decl: &FnDecl) -> RustType {
-    let return_ts_type = get_return_ts_type(ast_fnc_decl);
+    let return_ts_type = ts_ast::fn_decl::get_return_ts_type(ast_fnc_decl);
     ts_type_to_rust_type(&return_ts_type, &None)
 }
 
-pub fn get_return_ts_type(ast_fnc_decl: &FnDecl) -> TsType {
-    let ts_type_ann = ast_fnc_decl.function.return_type.as_ref();
-    let return_type_ann = ts_type_ann.clone().unwrap();
-    let return_type_ref = return_type_ann.type_ann.as_ts_type_ref().unwrap();
-    let return_type_params = return_type_ref.type_params.clone().unwrap();
-
-    let return_ts_type = *return_type_params.params[0].clone();
-    return_ts_type
-}
-
-pub fn get_param_ts_types(ast_fnc_decl: &FnDecl) -> Vec<TsType> {
-    let params = &ast_fnc_decl.function.params;
-    params.iter().fold(vec![], |acc, param| {
-        let param_type_ann = &param.pat.as_ident().unwrap().type_ann.as_ref();
-        let param_type_ann = param_type_ann.clone().unwrap();
-        let param_ts_type = *param_type_ann.type_ann.clone();
-
-        vec![acc, vec![param_ts_type]].concat()
-    })
-}
-
 pub fn generate_param_types(ast_fnc_decl: &FnDecl) -> Vec<RustType> {
-    get_param_ts_types(ast_fnc_decl)
+    ts_ast::fn_decl::get_param_ts_types(ast_fnc_decl)
         .iter()
         .map(|ts_type| ts_type_to_rust_type(ts_type, &None))
         .collect()
