+++ src/compiler/typescript_to_rust_redesign/azle_generate/src/ts_ast/fn_decl.rs
@@ -4,10 +4,7 @@ use swc_ecma_ast::{FnDecl, TsType, TsTypeAliasDecl};
 use syn::Ident;
 
 use super::{ts_type::get_dependent_types_for_ts_type, ts_type_alias_decl::generate_hash_map};
-use crate::{
-    azle_ast::CanisterMethodType,
-    generators::canister_methods::{get_param_ts_types, get_return_ts_type},
-};
+use crate::azle_ast::CanisterMethodType;
 
 pub fn get_canister_method_return_type(fn_decl: &FnDecl) -> Option<&TsType> {
     let ts_type = &*fn_decl.function.return_type.as_ref().unwrap().type_ann;
@@ -58,6 +55,27 @@ pub fn get_param_name_idents(fn_decl: &FnDecl) -> Vec<Ident> {
             )
         })
         .collect()
+}
+
+pub fn get_param_ts_types(ast_fnc_decl: &FnDecl) -> Vec<TsType> {
+    let params = &ast_fnc_decl.function.params;
+    params.iter().fold(vec![], |acc, param| {
+        let param_type_ann = &param.pat.as_ident().unwrap().type_ann.as_ref();
+        let param_type_ann = param_type_ann.clone().unwrap();
+        let param_ts_type = *param_type_ann.type_ann.clone();
+
+        vec![acc, vec![param_ts_type]].concat()
+    })
+}
+
+pub fn get_return_ts_type(ast_fnc_decl: &FnDecl) -> TsType {
+    let ts_type_ann = ast_fnc_decl.function.return_type.as_ref();
+    let return_type_ann = ts_type_ann.clone().unwrap();
+    let return_type_ref = return_type_ann.type_ann.as_ts_type_ref().unwrap();
+    let return_type_params = return_type_ref.type_params.clone().unwrap();
+
+    let return_ts_type = *return_type_params.params[0].clone();
+    return_ts_type
 }
 
 pub fn is_canister_method_type_fn_decl(
